+++ CHANGELOG.md
@@ -17,10 +17,16 @@ major/minor/micro version numbers like `05` (it'd have to be just `5`).
 ### Fixed
 ### Security
 
+## [2025.2.18]
+
+### Added
+* `Language.match` now supports case-insensitive matching
+  and ignores leading/trailing whitespace by default.
+
 ## [2025.2.8]
 
 ### Added
-* Explcitly indicated that the library is type-annotated.
+* Explicitly indicated that the library is type-annotated.
 
 ## [2025.1.28]